+++ packages/dso-toolkit/src/components/document-list/document-list.args.ts
@@ -1,4 +1,4 @@
-import { documentListContent } from "./document-list.content.js";
+import { documentListContent, documentListStickyContent } from "./document-list.content.js";
 import { DocumentList, DocumentListItemStatusDemoContent } from "./document-list.models.js";
 
 export function documentListMapper<TemplateFnReturnType>(
@@ -6,5 +6,13 @@ export function documentListMapper<TemplateFnReturnType>(
 ): DocumentList<TemplateFnReturnType> {
   return {
     items: documentListContent.items.map((c) => ({ ...c, status: demoMapper(c.status) })),
+  };
+}
+
+export function documentListStickyMapper<TemplateFnReturnType>(
+  demoMapper: ({ badge, date }: DocumentListItemStatusDemoContent) => TemplateFnReturnType
+): DocumentList<TemplateFnReturnType> {
+  return {
+    items: documentListStickyContent.items.map((c) => ({ ...c, status: demoMapper(c.status) })),
   };
 }