+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
-## Fixed
+### Added
+* sticky variant voor Document List Item Heading ([#1905](https://github.com/dso-toolkit/dso-toolkit/issues/1905))
+
+### Fixed
 * Tijdens release gaat deployen van website niet goed ([#1949](https://github.com/dso-toolkit/dso-toolkit/issues/1949)).
 * CDN verbeteringen ([#1952](https://github.com/dso-toolkit/dso-toolkit/issues/1952)).
 * Documentatie doorzoeken in Docusaurus ([#1950](https://github.com/dso-toolkit/dso-toolkit/issues/1950)).